requirements.txt
fastapi
uvicorn
litellm
python-dotenv
httpx
rich
pandas
numpy
scikit-learn
judgy
rank-bm25
tqdm
pydantic
typing-extensions
matplotlib
seaborn
plotly
python-fasthtml
monsterui
braintrust
